Lens.Micro.TH.Internal

Description

Functions used by Lens.Micro.TH. This is an internal module and it may go away or change at any time; do not depend on it.

Name utilities

class HasName t where

Has a Name

Methods

name :: Lens' t Name

Extract (or modify) the Name of something

newNames

Arguments

:: String
base name

-> Int
count

-> Q [Name]

Generate many new names from a given base name.

Type variable utilities

class HasTypeVars t where

Provides for the extraction of free type variables, and alpha renaming.

Methods

typeVarsEx :: Set Name -> Traversal' t Name

typeVars :: HasTypeVars t => Traversal' t Name

substTypeVars :: HasTypeVars t => Map Name Name -> t -> t

Miscellaneous utilities

datatypeTypeKinded :: DatatypeInfo -> Type

inlinePragma :: Name -> [DecQ]

Generate an INLINE pragma.

conAppsT :: Name -> [Type] -> Type

Apply arguments to a type constructor.

quantifyType :: Cxt -> Type -> Type

Template Haskell wants type variables declared in a forall, so we find all free type variables in a given type and declare them.

quantifyType' :: Set Name -> Cxt -> Type -> Type

This function works like quantifyType except that it takes a list of variables to exclude from quantification.

tvbToType :: TyVarBndr_ flag -> Type

Convert a TyVarBndr into its corresponding Type.

unSigT :: Type -> Type

Peel off a kind signature from a Type (if it has one).

Lens functions

elemOf :: Eq a => Getting (Endo [a]) s a -> a -> s -> Bool

lengthOf :: Getting (Endo [a]) s a -> s -> Int

setOf :: Ord a => Getting (Endo [a]) s a -> s -> Set a

_ForallT :: Traversal' Type ([TyVarBndrSpec], Cxt, Type)